+++ tests/test_checker/test_noqa.py
@@ -116,6 +116,7 @@ def test_noqa_fixture_disabled(absolute_path, all_violations):
         'Z445': 1,
         'Z446': 1,
         'Z447': 1,
+        'Z448': 1,
     }
 
     process = subprocess.Popen(